% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\alias{run}
\title{Compile experiment to html/javascript}
\usage{
make(experiment, directory = experiment@Title)

run(x, ...)
}
\description{
This function takes a \link{experiment} object and renders it into an html file which you open in your web browser.
}
