% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{make}
\alias{make}
\alias{run}
\title{Compile experiment to html/javascript}
\usage{
make(
  experiment,
  filename = str_to_title(str_remove_all(experiment@Title, "\\\\s*"))
)

run(x, ...)
}
\description{
This function takes a \link{experiment} object and renders it into an html file which you open in your web browser.
}
